解决分页错误的问题

+++ src/main/java/com/xinshijie/gallery/dto/UserVideoFindDto.java
@@ -31,7 +31,7 @@ public class UserVideoFindDto implements Serializable {
     private Integer isFree;
 
     @NotNull
-    private Long aid;
+    private Integer aid;
 
     private Long pageNum;
